edudesk_backend/requirements.txt
Django>=6.0
djangorestframework>=3.14
gunicorn>=23.0
djangorestframework-authtoken>=1.2
django-cors-headers>=4.3
google-genai>=1.0
python-docx>=1.1
Pillow>=10.0
django-ckeditor-5>=0.2.20
